+++ docs/changelog.md
@@ -7,8 +7,10 @@ DVR-Scan Changelog
 ### 1.4.1 (In Development)
 
 #### Changelog
+
  * Event start times are now correctly calculated when using `-fs`/`--frame-skip` ([#68](https://github.com/Breakthrough/DVR-Scan/issues/68), [#70](https://github.com/Breakthrough/DVR-Scan/issues/70))
     * Note that event start/end times may still be off by how many frames are skipped due to loss of context with respect to frame-accurate motion detection
+ * Event end times now take into account the number of skipped frames if `-fs`/`--frame-skip` if specified (e.g. it is assumed all skipped frames contained motion)
 
 ### 1.4 (February 8, 2022)